fix: Fix cartography issues
