+++ src/Sample.Shared.Views/Sample.Shared.Views.projitems
@@ -235,7 +235,11 @@
     <Content Include="$(MSBuildThisFileDirectory)Assets\Fonts\Guillon-SemiBold.ttf" />
     <Content Include="$(MSBuildThisFileDirectory)Assets\Fonts\segmdl2.ttf" />
     <Content Include="$(MSBuildThisFileDirectory)Assets\Pushpin\active.png" />
+    <Content Include="$(MSBuildThisFileDirectory)Assets\Pushpin\active.scale-200.png" />
+    <Content Include="$(MSBuildThisFileDirectory)Assets\Pushpin\active.scale-300.png" />
     <Content Include="$(MSBuildThisFileDirectory)Assets\Pushpin\inactive.png" />
+    <Content Include="$(MSBuildThisFileDirectory)Assets\Pushpin\inactive.scale-200.png" />
+    <Content Include="$(MSBuildThisFileDirectory)Assets\Pushpin\inactive.scale-300.png" />
     <Content Include="$(MSBuildThisFileDirectory)Assets\Pushpin\pin_bleu_normal_base.scale-200.png" />
     <Content Include="$(MSBuildThisFileDirectory)Assets\Pushpin\pin_bleu_normal_base.scale-300.png" />
     <Content Include="$(MSBuildThisFileDirectory)Assets\Pushpin\pin_bleu_normal_choisi.scale-200.png" />
